Refactor: Divide Component
